arivani/arivani/django_dynamic_url/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.34.14
botocore==1.34.14
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.8
django-allauth==0.60.1
django-cleanup==8.0.0
django-crispy-forms==2.1
django-import-export==3.3.7
django-sslserver==0.22
et-xmlfile==1.1.0
idna==3.6
jmespath==1.0.1
MarkupPy==1.14
numpy==1.24.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
pillow==10.2.0
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
